# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Fonoteca.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/library/scan": {
      "post": {
        "tags": [
          "Library"
        ],
        "summary": "Reconcile the catalogue's file list with the files on disk.",
        "description": "Walks the library root, adding files that are new, updating those whose size or modification time changed, and removing rows for files that are gone. Reads no file contents and modifies no file.",
        "operationId": "ScanLibrary",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryScanSummary"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "Whether a scan is running, and what the last one found.",
        "operationId": "GetLibraryScanStatus",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryScanStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/catalogue/artists": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Album artists in the library, or everyone credited on one.",
        "description": "Ordered by sort name, or by `sort=tracks` for the most-held first — ties keep the alphabetical order. `query` filters on the artist's name, case-insensitively, anywhere in the string. By default the list is the artists an album is billed to: the release's own credit line, collaborators included, plus anyone billed on every track of an album — which is how a conductor and an orchestra are found on a release the sleeve bills to the composer — and the track's own credit line for files no release has been attributed to yet. `scope=all` widens it to everyone the catalogue can reach a track through: composers and lyricists of the work, conductors, ensembles and guest features, which on a real library is an order of magnitude more names.",
        "operationId": "GetArtists",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalogue/artists/{id}": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "One artist and every track of theirs in the library.",
        "operationId": "GetArtist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/catalogue/releases/{id}/cover": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "The album's cover, as stored in the catalogue.",
        "description": "The first request for an album with no stored cover fetches the Cover Art Archive's front at 500px and keeps it, so the archive is asked once per album. Where the archive holds no front, Qobuz is asked for the same record — by barcode where there is one, by a matched title and artist otherwise — and refuses rather than guesses. An album neither has a picture of answers 404 and is remembered as such for a week, then asked about again, so a sleeve either gains later is still picked up. Served under an ETag with `no-cache`, so a changed cover shows on the next view and an unchanged one costs a 304.",
        "operationId": "GetReleaseCover",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Use an uploaded picture as the album's cover.",
        "description": "The request body is the image itself, its Content-Type one of JPEG, PNG, GIF, WebP, BMP or AVIF — never SVG, which is a document that runs script. At most 10 MB.",
        "operationId": "UploadReleaseCover",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtistDetailResponse": {
        "required": [
          "artist",
          "tracks",
          "discography",
          "profile"
        ],
        "type": "object",
        "properties": {
          "artist": {
            "$ref": "#/components/schemas/ArtistSummary"
          },
          "tracks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackRow"
            }
          },
          "discography": {
            "$ref": "#/components/schemas/ArtistDiscography"
          },
          "profile": {
            "$ref": "#/components/schemas/ArtistProfile"
          }
        }
      },
      "ArtistDiscography": {
        "required": [
          "known",
          "held",
          "fetchedAtUtc",
          "missing"
        ],
        "type": "object",
        "properties": {
          "known": {
            "type": "integer",
            "format": "int32"
          },
          "held": {
            "type": "integer",
            "format": "int32"
          },
          "fetchedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "missing": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiscographyRow"
            }
          }
        }
      },
...
      "ArtistListResponse": {
        "required": [
          "total",
          "items"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistSummary"
            }
          }
        }
      },
      "ArtistProfile": {
        "required": [
          "mbid",
          "name",
          "latinName",
          "portraitLookupUtc",
          "portraitUploaded",
          "banner",
          "bannerLookupUtc",
          "bannerUploaded",
          "biography",
          "biographyLookupUtc",
          "discographyLookupUtc",
          "edited",
          "members",
          "memberOf"
        ],
        "type": "object",
        "properties": {
          "mbid": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "latinName": {
            "type": [
              "null",
              "string"
            ]
          },
          "portraitLookupUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "portraitUploaded": {
            "type": "boolean"
          },
          "banner": {
            "type": [
              "null",
              "string"
            ]
          },
          "bannerLookupUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "bannerUploaded": {
            "type": "boolean"
          },
          "biography": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WrittenText"
              }
            ]
          },
          "biographyLookupUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "discographyLookupUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "edited": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelatedArtist"
            }
          },
          "memberOf": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelatedArtist"
            }
          }
        }
      },
      "ArtistSummary": {
        "required": [
          "id",
          "name",
          "sortName",
          "disambiguation",
          "type",
          "trackCount",
          "portrait",
          "country",
          "gender",
          "beganYear",
          "endedYear",
          "ended",
          "genres",
          "describedAtUtc",
          "following"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "sortName": {
            "type": [
              "null",
              "string"
            ]
          },
          "disambiguation": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "trackCount": {
            "type": "integer",
            "format": "int32"
          },
          "portrait": {
            "type": [
              "null",
              "string"
            ]
          },
          "country": {
            "type": [
              "null",
              "string"
            ]
          },
          "gender": {
            "type": [
              "null",
              "string"
            ]
          },
          "beganYear": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "endedYear": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "ended": {
            "type": "boolean"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "describedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "following": {
            "type": "boolean"
          }
        }
      },
...
      "LibraryScanStatusResponse": {
        "required": [
          "running",
          "lastCompleted"
        ],
        "type": "object",
        "properties": {
          "running": {
            "type": "boolean"
          },
          "lastCompleted": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LibraryScanSummary"
              }
            ]
          }
        }
      },
      "LibraryScanSummary": {
        "required": [
          "startedAtUtc",
          "completedAtUtc",
          "durationMilliseconds",
          "filesSeen",
          "added",
          "updated",
          "unchanged",
          "removed",
          "unreadableDirectories"
        ],
        "type": "object",
        "properties": {
          "startedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "completedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "durationMilliseconds": {
            "type": "integer",
            "format": "int64"
          },
          "filesSeen": {
            "type": "integer",
            "format": "int32"
          },
          "added": {
            "type": "integer",
            "format": "int32"
          },
          "updated": {
            "type": "integer",
            "format": "int32"
          },
          "unchanged": {
            "type": "integer",
            "format": "int32"
          },
          "removed": {
            "type": "integer",
            "format": "int32"
          },
          "unreadableDirectories": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RelatedArtist": {
        "required": [
          "id",
          "name",
          "portrait"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "portrait": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "TrackRow": {
        "required": [
          "recordingId",
          "title",
          "workTitle",
          "duration",
          "roles",
          "album",
          "folder",
          "files",
          "performers"
        ],
        "type": "object",
        "properties": {
          "recordingId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "workTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "duration": {
            "type": [
              "null",
              "string"
            ]
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "album": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TrackAlbum"
              }
            ]
          },
          "folder": {
            "type": "string"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileRow"
            }
          },
          "performers": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}